src/ingest/requirements.txt
azure-search-documents
azure-storage-blob
pypdf
tiktoken
python-dotenv